requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
py-notifier==0.1.0
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.0.1
python-dateutil==2.8.0
python-igraph==0.7.1.post6
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
wikipedia==1.4.0
